Refactor (#201)
* Refactor app structure

* Refactor imports

* Moved Data class to separate file

* Separated logging into model related logging and ml_tooling related logging

* Cleaned up imports and minor pep8

* Moved viz classes into submodule

* Set up __init__ imports

* Updated tests for text label positioning

* Fixed tests checking lowest value is using absolute lowest value

* Flake8 checks

* Updated Changelog
