+++ src/ml_tooling/data/__init__.py
@@ -1,0 +1,5 @@
+from .data import Data
+
+__all__ = [
+    'Data'
+]
