+++ src/ml_tooling/logging/__init__.py
@@ -1,0 +1,7 @@
+from .log_model import log_model
+from .logger import create_logger
+
+__all__ = [
+    'log_model',
+    'create_logger'
+]
