+++ src/ml_tooling/metrics/__init__.py
@@ -1,0 +1,11 @@
+from .confusion_matrix import confusion_matrix
+from .correlation import target_correlation, multi_collinearity
+from .lift_score import lift_score
+
+__all__ = [
+    'confusion_matrix',
+    'target_correlation',
+    'multi_collinearity',
+    'lift_score',
+    'permutation_importance'
+]
