+++ src/ml_tooling/plots/__init__.py
@@ -1,0 +1,17 @@
+from .confusion_matrix import plot_confusion_matrix
+from .correlation import plot_target_correlation
+from .feature_importance import plot_feature_importance
+from .lift_curve import plot_lift_curve
+from .prediction_error import plot_prediction_error
+from .residuals import plot_residuals
+from .roc_auc import plot_roc_auc
+
+__all__ = [
+    'plot_confusion_matrix',
+    'plot_target_correlation',
+    'plot_feature_importance',
+    'plot_lift_curve',
+    'plot_prediction_error',
+    'plot_residuals',
+    'plot_roc_auc'
+]
