+++ src/ml_tooling/result/__init__.py
@@ -1,0 +1,9 @@
+from .result import Result
+from .result_group import ResultGroup
+from .cv_result import CVResult
+
+__all__ = [
+    'Result',
+    'ResultGroup',
+    'CVResult'
+]
