+++ tests/test_metrics.py
@@ -2,14 +2,16 @@ import numpy as np
 import pandas as pd
 import pytest
 from sklearn.metrics import get_scorer
-from ml_tooling.utils import (MLToolingError,
-                              )
-from ml_tooling.metrics import (MetricError,
-                                lift_score,
+
+from ml_tooling.metrics import (lift_score,
                                 confusion_matrix,
-                                sorted_feature_importance,
-                                _permutation_importances,
+                                target_correlation
                                 )
+from ml_tooling.metrics.permutation_importance import _permutation_importances
+from ml_tooling.metrics.utils import (MetricError,
+                                      _sort_values,
+                                      )
+from ml_tooling.utils import MLToolingError
 
 
 class TestLiftScore:
@@ -50,7 +52,7 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Feature 2'])
         importance = np.array([.8, .7])
 
-        result_labels, result_importance = sorted_feature_importance(labels, importance)
+        result_labels, result_importance = _sort_values(labels, importance)
         assert np.all(labels == result_labels)
         assert np.all(importance == result_importance)
 
@@ -58,7 +60,7 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Features 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels, importance, top_n=2)
+        result_labels, result_importance = _sort_values(labels, importance, top_n=2)
         assert ['Feature 5', 'Feature 4'] == list(result_labels)
         assert [.5, .4] == list(result_importance)
 
@@ -66,7 +68,7 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Features 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels, importance, top_n=.2)
+        result_labels, result_importance = _sort_values(labels, importance, top_n=.2)
         assert ['Feature 5'] == list(result_labels)
         assert [.5] == list(result_importance)
 
@@ -74,7 +76,7 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Feature 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels, importance, bottom_n=2)
+        result_labels, result_importance = _sort_values(labels, importance, bottom_n=2)
         assert ['Feature 1', 'Feature 2'] == list(result_labels)
         assert [.1, .2] == list(result_importance)
 
@@ -82,8 +84,8 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Feature 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels, importance,
-                                                                     bottom_n=.2)
+        result_labels, result_importance = _sort_values(labels, importance,
+                                                        bottom_n=.2)
         assert ['Feature 1'] == list(result_labels)
         assert [.1] == list(result_importance)
 
@@ -91,10 +93,10 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Feature 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels,
-                                                                     importance,
-                                                                     bottom_n=2,
-                                                                     top_n=1)
+        result_labels, result_importance = _sort_values(labels,
+                                                        importance,
+                                                        bottom_n=2,
+                                                        top_n=1)
         assert ['Feature 5', 'Feature 1', 'Feature 2'] == list(result_labels)
         assert [.5, .1, .2] == list(result_importance)
 
@@ -102,12 +104,22 @@ class TestFeatureImportance:
         labels = np.array(['Feature 1', 'Feature 2', 'Feature 3', 'Feature 4', 'Feature 5'])
         importance = np.array([.1, .2, .3, .4, .5])
 
-        result_labels, result_importance = sorted_feature_importance(labels,
-                                                                     importance,
-                                                                     bottom_n=.2,
-                                                                     top_n=.2)
+        result_labels, result_importance = _sort_values(labels,
+                                                        importance,
+                                                        bottom_n=.2,
+                                                        top_n=.2)
         assert ['Feature 5', 'Feature 1'] == list(result_labels)
         assert [.5, .1] == list(result_importance)
+
+    def test_sorted_feature_importance_sorts_by_absolute_value_correctly(self):
+        labels = np.array(['Feature 1', 'Feature 2', 'Feature 3', 'Feature 4', 'Feature 5'])
+        importance = np.array([.1, .2, .3, -.4, -.5])
+
+        result_labels, result_importance = _sort_values(labels,
+                                                        importance,
+                                                        sort='abs')
+
+        assert np.all(result_importance == np.array([-.5, -.4, .3, .2, .1]))
 
     def test_permutation_importances_raises(self, regression):
         x = regression.data.train_x
@@ -179,3 +191,20 @@ class TestFeatureImportance:
 
         assert np.all(importance_parellel == importance_single)
         assert baseline_single == baseline_parellel
+
+
+class TestCorrelation:
+    def test_correlation_gives_expected_result(self):
+        x_values = pd.DataFrame({"col1": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]})
+        y_values = np.array([10, 20, 30, 40, 50, 60, 70, 80, 90])
+        corr = target_correlation(x_values, y_values)
+        assert corr.at["col1"] == 1.0
+
+    def test_anscombes_quartet_gives_expected_result(self):
+        x_values = pd.DataFrame(
+            {"col1": [8.04, 6.95, 7.58, 8.81, 8.33, 9.96, 7.24, 4.26, 10.84, 4.82, 5.68],
+             "col2": [9.14, 8.14, 8.74, 8.77, 9.26, 8.10, 6.13, 3.10, 9.13, 7.26, 4.74],
+             "col3": [7.46, 6.77, 12.74, 7.11, 7.81, 8.84, 6.08, 5.39, 8.15, 6.42, 5.73]})
+        y_values = np.array([10, 8, 13, 9, 11, 14, 6, 4, 12, 7, 5])
+        corr = target_correlation(x_values, y_values)
+        assert (corr.round(3) == 0.816).all()
