+++ tests/test_visualize.py
@@ -1,29 +1,28 @@
 """
 Test file for visualisations
 """
-import pytest
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
+import pytest
 from matplotlib.axes import Axes
 from sklearn.datasets import load_iris
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import roc_curve
 from sklearn.pipeline import Pipeline
+from sklearn.svm import SVC
 
 from ml_tooling import BaseClassModel
-from ml_tooling.transformers import ToCategorical
+from ml_tooling.metrics.permutation_importance import _get_feature_importance
+from ml_tooling.metrics.permutation_importance import _permutation_importances
 from ml_tooling.plots import (plot_lift_curve,
-                              VizError,
-                              _get_feature_importance,
                               plot_confusion_matrix,
                               )
-
-from ml_tooling.result import RegressionVisualize, ClassificationVisualize
-from ml_tooling.metrics import _permutation_importances
+from ml_tooling.plots.utils import VizError
+from ml_tooling.result.viz import RegressionVisualize, ClassificationVisualize
+from ml_tooling.transformers import ToCategorical
 from ml_tooling.utils import get_scoring_func
-from sklearn.svm import SVC
 
 
 class TestVisualize:
@@ -143,7 +142,7 @@ class TestFeatureImportancePlot:
                                                                                    classifier):
         ax = classifier.result.plot.feature_importance(bottom_n=.2, samples=100)
         assert 1 == len(ax.texts)
-        assert {'-0.05'} == {text._text for text in ax.texts}
+        assert {'0.03'} == {text._text for text in ax.texts}
         assert 'Feature Importance - LogisticRegression - Bottom 20%' == ax.title._text
         assert 'Features' == ax.get_ylabel()
         assert 'Importance:  Decrease in accuracy from baseline of 0.7' == ax.get_xlabel()
@@ -152,7 +151,7 @@ class TestFeatureImportancePlot:
     def test_feature_importance_plots_correct_if_top_n_is_int_and_bottom_n_is_int(self, classifier):
         ax = classifier.result.plot.feature_importance(top_n=1, bottom_n=1, samples=100)
         assert 2 == len(ax.texts)
-        assert {'0.12', '-0.05'} == {text._text for text in ax.texts}
+        assert {'0.12', '0.03'} == {text._text for text in ax.texts}
         assert 'Feature Importance - LogisticRegression - Top 1 - Bottom 1' == ax.title._text
         assert 'Features' == ax.get_ylabel()
         assert 'Importance:  Decrease in accuracy from baseline of 0.7' == ax.get_xlabel()
@@ -162,7 +161,7 @@ class TestFeatureImportancePlot:
                                                                                         classifier):
         ax = classifier.result.plot.feature_importance(top_n=1, bottom_n=.2, samples=100)
         assert 2 == len(ax.texts)
-        assert {'0.12', '-0.05'} == {text._text for text in ax.texts}
+        assert {'0.12', '0.03'} == {text._text for text in ax.texts}
         assert 'Feature Importance - LogisticRegression - Top 1 - Bottom 20%' == ax.title._text
         assert 'Features' == ax.get_ylabel()
         assert 'Importance:  Decrease in accuracy from baseline of 0.7' == ax.get_xlabel()
@@ -172,7 +171,7 @@ class TestFeatureImportancePlot:
                                                                                         classifier):
         ax = classifier.result.plot.feature_importance(top_n=.2, bottom_n=1, samples=100)
         assert 2 == len(ax.texts)
-        assert {'0.12', '-0.05'} == {text._text for text in ax.texts}
+        assert {'0.12', '0.03'} == {text._text for text in ax.texts}
         assert 'Feature Importance - LogisticRegression - Top 20% - Bottom 1' == ax.title._text
         assert 'Features' == ax.get_ylabel()
         assert 'Importance:  Decrease in accuracy from baseline of 0.7' == ax.get_xlabel()
